Renamed optimized to compact
